pyproject.toml
[tool.poetry]
name = "partysys"
version = "0.1.0"
description = ""
authors = ["FolkiDevv folkidevv.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
sentry-sdk = "^1.39.2"
dbutils = "^3.0.3"
pymysql = "^1.1.0"
cryptography = "^42.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"
ruff-lsp = "^0.0.53"

[tool.ruff]
# Set the maximum line length to 79.
line-length = 80

[tool.ruff.lint]
extend-select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "N",  #  pep8-naming
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
